requirements.txt
anyio==4.10.0
asgiref==3.9.1
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
crontab==1.0.5
Django==5.2.5
django-crontab==0.7.1
django-filter==25.1
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
multidict==6.6.4
promise==2.3
propcache==0.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
